requirements.txt
python-dotenv
imap-tools
requests
pykrx
opendartreader
